refactor(call): extract _TransferCoordinator extension from CallBloc (#1017)
Move all call transfer handlers (~190 lines) into a dedicated part file
(transfer_coordinator.dart) as a private extension.

Includes:
- _onCallControlEventBlindTransferInitiated
- _onCallControlEventAttendedTransferInitiated
- _onCallControlEventBlindTransferSubmitted
- _onCallControlEventAttendedTransferSubmitted
- _onCallControlEventAttendedRequestApproved
- _onCallControlEventAttendedRequestDeclined
